separating env variables in crates
