+++ cw-orch-daemon/src/lib.rs
@@ -14,6 +14,7 @@ pub mod state;
 pub mod sync;
 pub mod tx_resp;
 // expose these as mods as they can grow
+pub mod env;
 pub mod keys;
 pub mod live_mock;
 mod log;
@@ -23,6 +24,7 @@ pub mod tx_builder;
 pub use self::{builder::*, channel::*, core::*, error::*, state::*, sync::*, tx_resp::*};
 pub use cw_orch_networks::chain_info::*;
 pub use cw_orch_networks::networks;
+pub use env::DaemonEnvVars;
 pub use sender::Wallet;
 pub use tx_builder::TxBuilder;
