+++ cw-orch-daemon/src/state.rs
@@ -1,12 +1,11 @@
 use super::error::DaemonError;
-use crate::{channel::GrpcChannel, networks::ChainKind};
+use crate::{channel::GrpcChannel, env::default_state_folder, networks::ChainKind, DaemonEnvVars};
 
 use cosmwasm_std::Addr;
 use cw_orch_core::{
-    env::default_state_folder,
     environment::StateInterface,
     log::{connectivity_target, local_target},
-    CwEnvError, CwOrchEnvVars,
+    CwEnvError,
 };
 use ibc_chain_registry::chain::ChainData;
 use serde::Serialize;
@@ -115,7 +114,7 @@ impl DaemonState {
     /// Returns the path of the file where the state of `cw-orchestrator` is stored.
     pub fn state_file_path() -> Result<String, DaemonError> {
         // check if STATE_FILE en var is configured, default to state.json
-        let env_file_path = CwOrchEnvVars::load()?.state_file;
+        let env_file_path = DaemonEnvVars::load()?.state_file;
         let state_file_path = if env_file_path.is_relative() {
             // If it's relative, we check if it start with "."
             let first_path_component = env_file_path
@@ -236,9 +235,7 @@ impl StateInterface for DaemonState {
 pub mod test {
     use std::env;
 
-    use cw_orch_core::env::STATE_FILE_ENV_NAME;
-
-    use crate::DaemonState;
+    use crate::{env::STATE_FILE_ENV_NAME, DaemonState};
 
     #[test]
     fn test_env_variable_state_path() -> anyhow::Result<()> {
