+++ cw-orch-daemon/tests/common/mod.rs
@@ -7,7 +7,8 @@ mod node {
 
     use ctor::{ctor, dtor};
 
-    use cw_orch_core::CwOrchEnvVars;
+    use cw_orch_daemon::DaemonEnvVars;
+    use cw_orch_networks::NetworkEnvVars;
     use duct::cmd;
 
     // Config
@@ -153,7 +154,7 @@ mod node {
         }
         let image = env::var("JUNO_IMAGE").unwrap();
 
-        if CwOrchEnvVars::load().unwrap().local_mnemonic.is_none() {
+        if NetworkEnvVars::load().unwrap().local_mnemonic.is_none() {
             env::set_var("LOCAL_MNEMONIC", LOCAL_MNEMONIC);
         }
 
@@ -161,11 +162,11 @@ mod node {
         log::info!("Using CONTAINER_NAME: {}", container);
         log::info!(
             "Using STATE_FILE: {}",
-            CwOrchEnvVars::load().unwrap().state_file.display()
+            DaemonEnvVars::load().unwrap().state_file.display()
         );
         log::info!(
             "Using LOCAL_MNEMONIC: {:?}",
-            CwOrchEnvVars::load().unwrap().local_mnemonic
+            NetworkEnvVars::load().unwrap().local_mnemonic
         );
 
         container::start(&container, &image);
