+++ packages/cw-orch-core/src/contract/contract_instance.rs
@@ -1,10 +1,10 @@
 //! Main functional component for interacting with a contract. Used as the base for generating contract interfaces.
 use super::interface_traits::Uploadable;
 use crate::{
+    env::CoreEnvVars,
     environment::{ChainState, IndexResponse, StateInterface, TxHandler, TxResponse},
     error::CwEnvError,
     log::{contract_target, transaction_target},
-    CwOrchEnvVars,
 };
 
 use crate::environment::QueryHandler;
@@ -334,7 +334,7 @@ impl<Chain: ChainState + QueryHandler> Contract<Chain> {
 
 /// Helper to serialize objects (JSON or Rust DEBUG)
 fn log_serialize_message<E: Serialize + Debug>(msg: &E) -> Result<String, CwEnvError> {
-    if CwOrchEnvVars::load()?.serialize_json {
+    if CoreEnvVars::load()?.serialize_json {
         Ok(serde_json::to_string(msg)?)
     } else {
         Ok(format!("{:#?}", msg))
