+++ packages/cw-orch-networks/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod chain_info;
+pub mod env;
+pub use env::NetworkEnvVars;
 pub mod networks;
 
 pub use chain_info::*;
